+++ src/pocketmine/entity/Entity.php
@@ -31,6 +31,19 @@ use pocketmine\block\Portal;
 use pocketmine\block\Water;
 use pocketmine\block\SlimeBlock;
 use pocketmine\entity\Item as DroppedItem;
+use pocketmine\entity\Neutral\CaveSpider;
+use pocketmine\entity\Neutral\Enderman;
+use pocketmine\entity\Neutral\PigZombie;
+use pocketmine\entity\Neutral\Spider;
+use pocketmine\entity\Passive\Bat;
+use pocketmine\entity\Passive\Chicken;
+use pocketmine\entity\Passive\Cow;
+use pocketmine\entity\Passive\Mooshroom;
+use pocketmine\entity\Passive\Pig;
+use pocketmine\entity\Passive\Rabbit;
+use pocketmine\entity\Passive\Sheep;
+use pocketmine\entity\Passive\Squid;
+use pocketmine\entity\Passive\Villager;
 use pocketmine\event\entity\EntityDamageEvent;
 use pocketmine\event\entity\EntityDespawnEvent;
 use pocketmine\event\entity\EntityEffectAddEvent;
