+++ src/pocketmine/entity/Neutral/CaveSpider.php
@@ -19,29 +19,29 @@
  *
  */
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\Neutral;
 
+use pocketmine\entity\Monster;
 use pocketmine\network\protocol\AddEntityPacket;
 use pocketmine\Player;
-use pocketmine\event\entity\EntityDamageByEntityEvent;
-use pocketmine\item\Item as ItemItem;
 
-class Spider extends Monster{
-	const NETWORK_ID = 35;
-	public $width = 0.3;
-	public $length = 0.9;
-	public $height = 1.9;
+class CaveSpider extends Monster {
+	const NETWORK_ID = 40;
+
+	public $width = 1;
+	public $length = 1;
+	public $height = 0.5;
 
 	public $dropExp = [5, 5];
-	
-	public function getName() : string{
-		return "Spider";
-	}
 
+	public function getName() : string{
+		return "Cave Spider";
+	}
+	
 	public function spawnTo(Player $player){
 		$pk = new AddEntityPacket();
 		$pk->eid = $this->getId();
-		$pk->type = Spider::NETWORK_ID;
+		$pk->type = CaveSpider::NETWORK_ID;
 		$pk->x = $this->x;
 		$pk->y = $this->y;
 		$pk->z = $this->z;
@@ -53,27 +53,6 @@ class Spider extends Monster{
 		$pk->metadata = $this->dataProperties;
 		$player->dataPacket($pk);
 
-
 		parent::spawnTo($player);
-	}
-	
-	public function getDrops(){
-		$drops = array(ItemItem::get(ItemItem::STRING, 0, 1));
-		if ($this->lastDamageCause instanceof EntityDamageByEntityEvent and $this->lastDamageCause->getEntity() instanceof Player) {
-			if (mt_rand(0, 199) < 5) {
-				switch (mt_rand(0, 2)) {
-					case 0:
-						$drops[] = ItemItem::get(ItemItem::IRON_INGOT, 0, 1);
-						break;
-					case 1:
-						$drops[] = ItemItem::get(ItemItem::CARROT, 0, 1);
-						break;
-					case 2:
-						$drops[] = ItemItem::get(ItemItem::POTATO, 0, 1);
-						break;
-				}
-			}
-		}
-		return $drops;
 	}
 }
