+++ src/pocketmine/entity/Passive/Cow.php
@@ -19,31 +19,32 @@
  *
  */
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\Passive;
 
+use pocketmine\entity\Animal;
 use pocketmine\item\enchantment\Enchantment;
 use pocketmine\network\protocol\AddEntityPacket;
 use pocketmine\Player;
 use pocketmine\event\entity\EntityDamageByEntityEvent;
 use pocketmine\item\Item as ItemItem;
 
-class Pig extends Animal{
-	const NETWORK_ID = 12;
+class Cow extends Animal {
+	const NETWORK_ID = 11;
 
 	public $width = 0.3;
 	public $length = 0.9;
-	public $height = 1.9;
+	public $height = 1.8;
 
 	public $dropExp = [1, 3];
 	
 	public function getName() : string{
-		return "Pig";
+		return "Cow";
 	}
 	
 	public function spawnTo(Player $player){
 		$pk = new AddEntityPacket();
 		$pk->eid = $this->getId();
-		$pk->type = Pig::NETWORK_ID;
+		$pk->type = Cow::NETWORK_ID;
 		$pk->x = $this->x;
 		$pk->y = $this->y;
 		$pk->z = $this->z;
@@ -64,7 +65,12 @@ class Pig extends Animal{
 		if($cause instanceof EntityDamageByEntityEvent and $cause->getDamager() instanceof Player){
 			$lootingL = $cause->getDamager()->getItemInHand()->getEnchantmentLevel(Enchantment::TYPE_WEAPON_LOOTING);
 		}
-		$drops = array(ItemItem::get(ItemItem::RAW_PORKCHOP, 0, mt_rand(1, 3 + $lootingL)));
+		$drops = array(ItemItem::get(ItemItem::RAW_BEEF, 0, mt_rand(1, 3 + $lootingL)));
+		$drops[] = ItemItem::get(ItemItem::LEATHER, 0, mt_rand(0, 2 + $lootingL));
+		//TODO: add judgement for Steak
+		/*if ($this->lastDamageCause instanceof EntityDamageByEntityEvent and $this->lastDamageCause->getEntity() instanceof Player) {
+			$drops[] = ItemItem::get(ItemItem::LEATHER, 0, mt_rand(0,2));
+		}*/
 		return $drops;
 	}
 }
