+++ src/pocketmine/entity/Passive/Mooshroom.php
@@ -10,7 +10,7 @@
  * \_____/ |_____| |_|  \_| |_| /_____/  /_/     /_____/
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
+ * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
@@ -19,28 +19,30 @@
  *
  */
 
-namespace pocketmine\entity;
+namespace pocketmine\entity\Passive;
 
+use pocketmine\entity\Animal;
+use pocketmine\item\enchantment\Enchantment;
 use pocketmine\network\protocol\AddEntityPacket;
 use pocketmine\Player;
+use pocketmine\event\entity\EntityDamageByEntityEvent;
+use pocketmine\item\Item as ItemItem;
 
-class CaveSpider extends Monster{
-	const NETWORK_ID = 40;
+class Mooshroom extends Animal {
+	const NETWORK_ID = 16;
 
-	public $width = 1;
-	public $length = 1;
-	public $height = 0.5;
-
-	public $dropExp = [5, 5];
-
+	public $width = 0.3;
+	public $length = 0.9;
+	public $height = 1.8;
+	
 	public function getName() : string{
-		return "Cave Spider";
+		return "Mooshroom";
 	}
 	
 	public function spawnTo(Player $player){
 		$pk = new AddEntityPacket();
 		$pk->eid = $this->getId();
-		$pk->type = CaveSpider::NETWORK_ID;
+		$pk->type = Mooshroom::NETWORK_ID;
 		$pk->x = $this->x;
 		$pk->y = $this->y;
 		$pk->z = $this->z;
@@ -53,5 +55,16 @@ class CaveSpider extends Monster{
 		$player->dataPacket($pk);
 
 		parent::spawnTo($player);
+	}
+	
+	public function getDrops(){
+		$lootingL = 0;
+		$cause = $this->lastDamageCause;
+		if($cause instanceof EntityDamageByEntityEvent and $cause->getDamager() instanceof Player){
+			$lootingL = $cause->getDamager()->getItemInHand()->getEnchantmentLevel(Enchantment::TYPE_WEAPON_LOOTING);
+		}
+		$drops = array(ItemItem::get(ItemItem::RAW_BEEF, 0, mt_rand(1, 3 + $lootingL)));
+		$drops[] = ItemItem::get(ItemItem::LEATHER, 0, mt_rand(0, 2 + $lootingL));
+		return $drops;
 	}
 }
